requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
Werkzeug==2.3.7
